# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Прогноз отказов оборудования коллекторов",
    "description": "REST API сервиса прогноза отказов датчиков коллекторов на 24 часа (ЛЦТ-2026).\n\n**Вход:** нажмите *Authorize* и введите логин и пароль (в демо-режиме — демо-пользователя из README), поля client_id/secret оставьте пустыми.\n\n**Машина времени:** все запросы на чтение принимают `at`. Демо-данные покрывают 01.07.2026–31.08.2026, например `at=2026-08-01T12:00:00`.\n\n**Симуляция потока** (`/api/sim/*`): проигрывание суток по заранее рассчитанным прогнозам с ускорением; пока она идёт, запросы без `at` работают на модельное время. Онлайн-пересчёт моделью не выполняется.\n\nВремя — московское, без часового пояса.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/predictions/{prediction_id}/recommendation": {
      "get": {
        "tags": [
          "Прогнозы"
        ],
        "summary": "Рекомендация по ТО для прогноза (прозрачные правила, не ML)",
        "description": "Действие, обоснование со ссылкой на факты, приоритет, срок, исполнитель и «чего не делать». Факты берутся на момент среза прогноза: причины прогноза, отказы канала за 30/90 дней, соседи по объекту, питание объекта, заявки по датчику, точность модели по типу датчика. Правила — `backend/app/recommendations/rules.yaml` (проект правил, требует согласования со специалистами эксплуатации).",
        "operationId": "get_recommendation_api_predictions__prediction_id__recommendation_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prediction Id"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceAdvice"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/work-orders": {
      "post": {
        "tags": [
          "Заявки"
        ],
        "summary": "Создать черновик заявки (из прогноза — с автозаполнением)",
        "description": "Из `prediction_id` подставляются датчик, объект и — по рекомендации движка правил ТО (`GET /api/predictions/{id}/recommendation`) — приоритет, срок, исполнитель, рекомендация и описание с обоснованием и причинами прогноза. Любое поле можно передать явно. `at` в запросе — момент «машины времени», которым датируется заявка.",
        "operationId": "create_work_order_api_work_orders_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkOrderIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Заявки"
        ],
        "summary": "Список заявок с фильтрами",
        "operationId": "list_work_orders_api_work_orders_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "draft",
                      "submitted",
                      "in_progress",
                      "done",
                      "cancelled"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Можно несколько",
              "title": "Status"
            },
            "description": "Можно несколько"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "critical"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Объект; включает все вложенные",
              "title": "Object Id"
            },
            "description": "Объект; включает все вложенные"
          },
          {
            "name": "channel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel Id"
            }
          },
          {
            "name": "prediction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prediction Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по номеру, датчику, исполнителю",
              "title": "Q"
            },
            "description": "Поиск по номеру, датчику, исполнителю"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 50,
              "title": "Limit"
            },
            "description": "Размер страницы"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Сдвиг от начала",
              "default": 0,
              "title": "Offset"
            },
            "description": "Сдвиг от начала"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Прогноз не найден"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaintenanceAdvice": {
        "properties": {
          "prediction_id": {
            "type": "integer",
            "title": "Prediction Id"
          },
          "channel_id": {
            "type": "integer",
            "title": "Channel Id"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At",
            "description": "Срез прогноза: все факты взяты на этот момент"
          },
          "rule_id": {
            "type": "string",
            "title": "Rule Id",
            "examples": [
              "mass_link_loss"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Массовая потеря связи в объекте"
            ]
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Что сделать"
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Обоснование простыми словами со ссылкой на факты и вывод"
          },
          "facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Facts",
            "description": "Факты обоснования по отдельности"
          },
          "conclusion": {
            "type": "string",
            "title": "Conclusion"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Priority"
          },
          "priority_label": {
            "type": "string",
            "title": "Priority Label"
          },
          "due_hours": {
            "type": "number",
            "title": "Due Hours",
            "description": "Рекомендуемый срок от среза прогноза, часов"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "assignee": {
            "type": "string",
            "title": "Assignee",
            "description": "Кому: служба или бригада"
          },
          "avoid": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoid",
            "description": "Чего не делать"
          },
          "fault_kind": {
            "type": "string",
            "enum": [
              "link",
              "fault",
              "disconnected",
              "value",
              "group",
              "power",
              "drift",
              "unknown"
            ],
            "title": "Fault Kind",
            "description": "Предполагаемый вид отказа"
          },
          "fault_kind_label": {
            "type": "string",
            "title": "Fault Kind Label"
          },
          "fault_kind_basis": {
            "type": "string",
            "title": "Fault Kind Basis",
            "description": "На чём основан вид отказа"
          },
          "plan": {
            "type": "boolean",
            "title": "Plan",
            "description": "Профилактическая работа — попадает во вкладку «Рекомендованные работы»"
          },
          "work_order_needed": {
            "type": "boolean",
            "title": "Work Order Needed",
            "description": "false — заявка не нужна (норма, наблюдение)"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id",
            "description": "Строка справочника recommendations (code = rule:<rule_id>)"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "examples": [
              "Проект правил, требует согласования со специалистами эксплуатации"
            ]
          },
          "rules_version": {
            "type": "string",
            "title": "Rules Version"
          }
        },
        "type": "object",
        "required": [
          "prediction_id",
          "channel_id",
          "at",
          "rule_id",
          "title",
          "action",
          "reason",
          "facts",
          "conclusion",
          "priority",
          "priority_label",
          "due_hours",
          "due_at",
          "assignee",
          "fault_kind",
          "fault_kind_label",
          "fault_kind_basis",
          "plan",
          "work_order_needed",
          "source",
          "rules_version"
        ],
        "title": "MaintenanceAdvice",
        "description": "Рекомендация по техническому обслуживанию: прозрачные правила (backend/app/recommendations/rules.yaml),\nне ML. Проект правил — требует согласования со специалистами эксплуатации."
      },
...
      "WorkOrderIn": {
        "properties": {
          "prediction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Id",
            "description": "Из какого прогноза создаётся черновик — поля заполнятся сами"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id",
            "description": "Нужен, если заявка не из прогноза"
          },
          "reason_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Id"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          },
          "recommendation_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Text"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "critical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "По умолчанию — из уровня риска"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At",
            "description": "По умолчанию — из приоритета"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          }
        },
        "type": "object",
        "title": "WorkOrderIn"
      },
      "WorkOrderList": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "at",
          "total",
          "items"
        ],
        "title": "WorkOrderList"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}